using int64 instead of uint

+++ objects/nodesStatistics.go
@@ -2,7 +2,7 @@ package objects
 
 // NodesStatistics is the structure for overall nodes statistics
 type NodesStatistics struct {
-	EmtpyNodesCount          uint `json:"emtpyNodesCount"`
-	SynchronizedNodesCount   uint `json:"synchronizedNodesCount"`
-	UnsynchronizedNodesCount uint `json:"unsynchronizedNodesCount"`
+	EmtpyNodesCount          int64 `json:"emtpyNodesCount"`
+	SynchronizedNodesCount   int64 `json:"synchronizedNodesCount"`
+	UnsynchronizedNodesCount int64 `json:"unsynchronizedNodesCount"`
 }
